proj/src/requirements.txt
dj-database-url==0.4.2
Django==1.11.3
django-crispy-forms==1.6.1
djangorestframework==3.6.3
gunicorn==19.7.1
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
psycopg2==2.7.1
pytz==2017.2
